Implements new trello label api
The number of labels attached to a board or card is no longer fixed. The
trello web api supports a few new calls to read, set and update those
labels. This commit reflects these changes in the classes:

- `Trello::Label`
- `Trello::Card`
- `Trello::Board`

These changes are brake with the old api `labels` in `Trello::Board`.
The `labels` method will return a list of `Trello::Label` instead of
`Trello::LabelName`. For a list of `Trello::LabelName` instances use
`label_names`.