Create a copy of a checklist
The copied checklist looks like the original but is a different
Trello checklist.  The checklist name and board will be the same,
and the new checklist will have items that look like the items
of the original checklist.

On Trello, I have created one card with a long checklist.  I want
to be able to create a hundred or so other cards on the board with
different names and descriptions but with the same checklist.
This allows me to create a copy of my card's checklist on Trello and
then give it to a new card.