Use included array of labels for Card#labels
The Trello API returns a JSON array of N Label objects as part of the Card
object. This change maps those to an array of `Trello::Label` objects,
saving N API calls.

Updated tests to account for card.labels changes

Updated the spec tests to include an array of labels for initializing
the `Card` object. Also modified the labels tests in `card_spec.rb` to
no longer expect `client.get` for `card.labels`. Added tests for the
`labels` attribute and the `card_labels` attribute.