add Card#move_to_list_on_any_board
Adds an ``move_to_list_on_any_board`` instance method to
``Trello::Card`` which may move the card to the specified list on any
board. It uses either ``move_to_list`` or ``move_to_board`` as the API
does not move the card if you use the wrong call, e.g. it does not
move the card when using ``move_to_board`` with the target list and
the card being on the same board and ``move_to_list`` does not move the
card if the list is on another board.