Fix equality of BasicData
There is a slight bug in the current implementation of `BasicData#==`:
it doesn't check for the concrete type of BasicData object, which allows
for a situation where a card could be identified as being equal to a
list (or any other subclass of `BasicData`).

This commit fixes the equality by also ensuring that the types of two
`BasicData` objects are the same.