Remove some versions of ruby 2.1 from .travis.yml
These don't appear to be [working in the build][1]. It looks like
TravisCI is trying to pull down a binary for them but they are not
available. They shouldn't be necessary to test against either. Ruby 2.3
is EOL, so support for that and prior versions could be removed in the
next major release.

[1]: https://travis-ci.org/jeremytregunna/ruby-trello/builds/597160203