Fix truncation issue in CI configuration entries
Zero terminated float values in GitHub Actions CI configurations are truncated, leading to unexpected behavior.  In this instance this results loading an unexpected Ruby or ActiveModel version for several CI matrix entries.

Specifically, the unquoted '3.0' is interpreted as '3', and the latest Ruby 3 version - at this time of writing 3.1.1 - is loaded.  To ensure that a 3.0.x version is loaded, we need to add quotes.

In the ActiveModel case, use of an unquoted '6.0' is interpreted as '6' and ActiveModel 6.1.4.6 is loaded at this time.  To ensure a 6.0.x version is loaded, we need to add quotes.

The unquoted '7.0' would have similar problems if and when a Rails 7.1.x is released..